+++ engage/hamls/includes/org.haml
@@ -1,7 +1,7 @@
 -load smartmin i18n humanize temba compress
 
 #org-area
-  {% if user_orgs|length > 0 %}
+  {% if user_org %}
     #btn-org-home
       -#emoji house
   {% endif %}
